rename loki.stdout to loki.echo
