+++ CHANGELOG.md
@@ -30,7 +30,7 @@ Main (unreleased)
 
   - `loki.source.docker` reads logs from Docker containers and forwards them to
     other `loki` components. (@tpaschalis)
-  - `loki.stdout` writes received logs to stdout. (@tpaschalis, @rfratto)
+  - `loki.echo` writes received logs to stdout. (@tpaschalis, @rfratto)
   - `otelcol.processor.tail_sampling` samples traces based on a set of defined
     policies from `otelcol` components before forwarding them to other
     `otelcol` components. (@erikbaranowski)
